requirements.txt
PyGithub>=2.3.0
python-dotenv>=1.0.1
gitpython>=3.1.31
gradio>=3.40
flake8>=6.0.0
pylint>=3.0.0
langchain>=0.0.300
langchain-community>=0.0.70
